NetKVM: BZ#1096505: Compiling of all windows target under debug

+++ NetKVM/wlh/ParaNdis6-Impl.cpp
@@ -303,7 +303,7 @@ static BOOLEAN MiniportMSIInterrupt(
         *QueueDefaultInterruptDpc = TRUE;
     }
 #else
-    *TargetProcessors = path->DPCTargetProcessor;
+    *TargetProcessors = (ULONG)path->DPCTargetProcessor;
     *QueueDefaultInterruptDpc = TRUE;
 #endif
